Refactor as PG-specific plugin
